added copy range to address, replace markers, and allow more single tests run, and keep test app running when single test(s) were chosen

+++ extension/src/frame.cpp
@@ -466,7 +466,7 @@ bool wxExFrame::UpdateStatusBar(wxExSTC* stc, const wxString& pane)
       {
         if (stc->SelectionIsRectangle())
         {
-          text = wxString::Format("%d,%d,%d", line, pos, stc->GetSelectedText().length());
+          text = wxString::Format("%d,%d,%d", line, pos, (int)stc->GetSelectedText().length());
         }
         else
         {
